move types from payload to eth2 crate
