+++ beacon_node/execution_layer/src/test_utils/execution_block_generator.rs
@@ -8,6 +8,7 @@ use crate::{
     },
     static_valid_tx, ExecutionBlockWithTransactions,
 };
+use eth2::types::BlobsBundle;
 use kzg::Kzg;
 use rand::thread_rng;
 use serde::{Deserialize, Serialize};
@@ -16,9 +17,9 @@ use std::sync::Arc;
 use tree_hash::TreeHash;
 use tree_hash_derive::TreeHash;
 use types::{
-    BlobSidecar, BlobsBundle, ChainSpec, EthSpec, ExecutionBlockHash, ExecutionPayload,
-    ExecutionPayloadCapella, ExecutionPayloadDeneb, ExecutionPayloadHeader, ExecutionPayloadMerge,
-    ForkName, Hash256, Transactions, Uint256,
+    BlobSidecar, ChainSpec, EthSpec, ExecutionBlockHash, ExecutionPayload, ExecutionPayloadCapella,
+    ExecutionPayloadDeneb, ExecutionPayloadHeader, ExecutionPayloadMerge, ForkName, Hash256,
+    Transactions, Uint256,
 };
 
 use super::DEFAULT_TERMINAL_BLOCK;
