+++ beacon_node/network/Cargo.toml
@@ -11,6 +11,7 @@ matches = "0.1.8"
 exit-future = "0.2.0"
 slog-term = "2.6.0"
 slog-async = "2.5.0"
+eth2 = {path="../../common/eth2"}
 
 [dependencies]
 beacon_chain =  { path = "../beacon_chain" }
