+++ consensus/types/src/payload.rs
@@ -1,9 +1,7 @@
-use crate::beacon_block_body::KzgCommitments;
 use crate::{test_utils::TestRandom, *};
 use derivative::Derivative;
 use serde::de::DeserializeOwned;
-use serde::{Deserialize, Deserializer, Serialize};
-use serde_json::Value;
+use serde::{Deserialize, Serialize};
 use ssz::{Decode, Encode};
 use ssz_derive::{Decode, Encode};
 use std::borrow::Cow;
@@ -970,85 +968,4 @@ impl<T: EthSpec> From<BlindedPayload<T>> for ExecutionPayloadHeader<T> {
             }
         }
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(untagged)]
-#[serde(bound = "E: EthSpec")]
-pub enum FullPayloadContents<E: EthSpec> {
-    Payload(ExecutionPayload<E>),
-    PayloadAndBlobs(ExecutionPayloadAndBlobs<E>),
-}
-
-impl<E: EthSpec> FullPayloadContents<E> {
-    pub fn new(
-        execution_payload: ExecutionPayload<E>,
-        maybe_blobs: Option<BlobsBundle<E>>,
-    ) -> Self {
-        match maybe_blobs {
-            None => Self::Payload(execution_payload),
-            Some(blobs_bundle) => Self::PayloadAndBlobs(ExecutionPayloadAndBlobs {
-                execution_payload,
-                blobs_bundle,
-            }),
-        }
-    }
-
-    pub fn payload_ref(&self) -> &ExecutionPayload<E> {
-        match self {
-            FullPayloadContents::Payload(payload) => payload,
-            FullPayloadContents::PayloadAndBlobs(payload_and_blobs) => {
-                &payload_and_blobs.execution_payload
-            }
-        }
-    }
-
-    pub fn block_hash(&self) -> ExecutionBlockHash {
-        self.payload_ref().block_hash()
-    }
-
-    pub fn deconstruct(self) -> (ExecutionPayload<E>, Option<BlobsBundle<E>>) {
-        match self {
-            FullPayloadContents::Payload(payload) => (payload, None),
-            FullPayloadContents::PayloadAndBlobs(payload_and_blobs) => (
-                payload_and_blobs.execution_payload,
-                Some(payload_and_blobs.blobs_bundle),
-            ),
-        }
-    }
-}
-
-impl<E: EthSpec> ForkVersionDeserialize for FullPayloadContents<E> {
-    fn deserialize_by_fork<'de, D: Deserializer<'de>>(
-        value: Value,
-        fork_name: ForkName,
-    ) -> Result<Self, D::Error> {
-        match fork_name {
-            ForkName::Merge | ForkName::Capella => serde_json::from_value(value)
-                .map(Self::Payload)
-                .map_err(serde::de::Error::custom),
-            ForkName::Deneb => serde_json::from_value(value)
-                .map(Self::PayloadAndBlobs)
-                .map_err(serde::de::Error::custom),
-            ForkName::Base | ForkName::Altair => Err(serde::de::Error::custom(format!(
-                "FullPayloadContents deserialization for {fork_name} not implemented"
-            ))),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(bound = "E: EthSpec")]
-pub struct ExecutionPayloadAndBlobs<E: EthSpec> {
-    pub execution_payload: ExecutionPayload<E>,
-    pub blobs_bundle: BlobsBundle<E>,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-#[serde(bound = "E: EthSpec")]
-pub struct BlobsBundle<E: EthSpec> {
-    pub commitments: KzgCommitments<E>,
-    pub proofs: KzgProofs<E>,
-    #[serde(with = "ssz_types::serde_utils::list_of_hex_fixed_vec")]
-    pub blobs: BlobsList<E>,
 }
