+++ consensus/types/src/signed_blob.rs
@@ -1,6 +1,7 @@
+use crate::sidecar::Sidecar;
 use crate::{
-    BlindedBlobSidecar, Blob, BlobSidecar, ChainSpec, Domain, EthSpec, Fork,
-    Hash256, Signature, SignedRoot, SigningData, test_utils::TestRandom,
+    test_utils::TestRandom, BlindedBlobSidecar, Blob, BlobSidecar, ChainSpec, Domain, EthSpec,
+    Fork, Hash256, Signature, SignedRoot, SigningData,
 };
 use bls::PublicKey;
 use derivative::Derivative;
@@ -12,7 +13,6 @@ use std::sync::Arc;
 use test_random_derive::TestRandom;
 use tree_hash::TreeHash;
 use tree_hash_derive::TreeHash;
-use crate::sidecar::Sidecar;
 
 #[derive(
     Debug,
